feat: added Users Cog
